added

+++ factor.h
@@ -1,5 +1,6 @@
 #ifndef FACTOR_h
 #define FACTOR_h
 #include <stdio.h>
-void factorize(long int number);
+#include <stdint.h>
+void factorize(unsigned long long number);
 #endif
